pyproject.toml
[tool.poetry]
name = "silhouette"
version = "0.1.0"
description = ""
authors = [
    "sazzeck <sazzeckprog@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
Django = "^4.1"
djangorestframework = "^3.13.1"
aiogram = "^2.22.1"
aiohttp = "^3.8.1"
django-redis = "^5.2.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
